refactor: extract best transactions
